AG-7369 - Refactor tooltip handling out of chart.ts.
