+++ charts-packages/ag-charts-community/src/chart/themes/chartTheme.ts
@@ -13,8 +13,8 @@ import {
     AgCartesianSeriesTheme,
     AgHierarchySeriesTheme,
 } from '../agChartOptions';
-import { Chart } from '../chart';
 import { TimeInterval } from '../../util/time/interval';
+import { DEFAULT_TOOLTIP_CLASS } from '../tooltip/tooltip';
 
 const palette: AgChartThemePalette = {
     fills: ['#f3622d', '#fba71b', '#57b757', '#41a9c9', '#4258c9', '#9a42c8', '#c84164', '#888888'],
@@ -247,7 +247,7 @@ export class ChartTheme {
                 enabled: true,
                 tracking: true,
                 delay: 0,
-                class: Chart.defaultTooltipClass,
+                class: DEFAULT_TOOLTIP_CLASS,
             },
         };
     }
